Split out http and mongo functionality into distinct components
