+++ chapter6/api/main.go
@@ -1,11 +1,11 @@
 package main
 
 import (
-	"context"
 	"flag"
-	"gopkg.in/mgo.v2"
 	"log"
 	"net/http"
+
+	"gopkg.in/mgo.v2"
 )
 
 func main() {
@@ -19,56 +19,10 @@ func main() {
 		log.Fatalln("failed to connect to mongo:", err)
 	}
 	defer db.Close()
-	s := &Server{
-		db: db,
-	}
+	s := NewHttpServer(NewMongoStorage(db))
 	mux := http.NewServeMux()
-	mux.HandleFunc("/polls/", withCORS(withAPIKey(s.handlePolls)))
+	mux.HandleFunc("/polls/", s.Routes)
 	log.Println("Starting web server on", *addr)
 	http.ListenAndServe(":8080", mux)
 	log.Println("Stopping...")
-}
-
-// Server is the API server.
-type Server struct {
-	db *mgo.Session
-}
-
-func withCORS(fn http.HandlerFunc) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Access-Control-Allow-Origin", "*")
-		w.Header().Set("Access-Control-Expose-Headers", "Location")
-		fn(w, r)
-	}
-}
-
-type contextKey struct {
-	name string
-}
-
-var contextKeyAPIKey = &contextKey{"api-key"}
-
-func APIKey(ctx context.Context) (string, bool) {
-	key := ctx.Value(contextKeyAPIKey)
-	if key == nil {
-		return "", false
-	}
-	keystr, ok := key.(string)
-	return keystr, ok
-}
-
-func withAPIKey(fn http.HandlerFunc) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		key := r.URL.Query().Get("key")
-		if !isValidAPIKey(key) {
-			respondErr(w, r, http.StatusUnauthorized, "invalid API key")
-			return
-		}
-		ctx := context.WithValue(r.Context(), contextKeyAPIKey, key)
-		fn(w, r.WithContext(ctx))
-	}
-}
-
-func isValidAPIKey(key string) bool {
-	return key == "abc123"
 }
