+++ chapter6/api/polls.go
@@ -1,13 +1,19 @@
 package main
 
 import (
-	"net/http"
+	"errors"
 
-	"gopkg.in/mgo.v2"
 	"gopkg.in/mgo.v2/bson"
 )
 
-type poll struct {
+var (
+	// ErrCannotDeleteAll is used to identify a client-facing
+	// error when deleting a poll from storage
+	ErrCannotDeleteAll = errors.New("cannot delete all polls")
+)
+
+// Poll represents a poll. ID should not be a BSON object
+type Poll struct {
 	ID      bson.ObjectId  `bson:"_id" json:"id"`
 	Title   string         `json:"title" bson:"title"`
 	Options []string       `json:"options"`
@@ -15,81 +21,10 @@ type poll struct {
 	APIKey  string         `json:"apikey"`
 }
 
-func (s *Server) handlePolls(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		s.handlePollsGet(w, r)
-		return
-	case "POST":
-		s.handlePollsPost(w, r)
-		return
-	case "DELETE":
-		s.handlePollsDelete(w, r)
-		return
-	case "OPTIONS":
-		w.Header().Set("Access-Control-Allow-Methods", "DELETE")
-		respond(w, r, http.StatusOK, nil)
-		return
-	}
-	// not found
-	respondHTTPErr(w, r, http.StatusNotFound)
-}
-
-func (s *Server) handlePollsGet(w http.ResponseWriter, r *http.Request) {
-	session := s.db.Copy()
-	defer session.Close()
-	c := session.DB("ballots").C("polls")
-	var q *mgo.Query
-	p := NewPath(r.URL.Path)
-	if p.HasID() {
-		// get specific poll
-		q = c.FindId(bson.ObjectIdHex(p.ID))
-	} else {
-		// get all polls
-		q = c.Find(nil)
-	}
-	var result []*poll
-	if err := q.All(&result); err != nil {
-		respondErr(w, r, http.StatusInternalServerError, err)
-		return
-	}
-	respond(w, r, http.StatusOK, &result)
-}
-
-func (s *Server) handlePollsPost(w http.ResponseWriter, r *http.Request) {
-	session := s.db.Copy()
-	defer session.Close()
-	c := session.DB("ballots").C("polls")
-	var p poll
-	if err := decodeBody(r, &p); err != nil {
-		respondErr(w, r, http.StatusBadRequest, "failed to read poll from request", err)
-		return
-	}
-	apikey, ok := APIKey(r.Context())
-	if ok {
-		p.APIKey = apikey
-	}
-	p.ID = bson.NewObjectId()
-	if err := c.Insert(p); err != nil {
-		respondErr(w, r, http.StatusInternalServerError, "failed to insert poll", err)
-		return
-	}
-	w.Header().Set("Location", "polls/"+p.ID.Hex())
-	respond(w, r, http.StatusCreated, nil)
-}
-
-func (s *Server) handlePollsDelete(w http.ResponseWriter, r *http.Request) {
-	session := s.db.Copy()
-	defer session.Close()
-	c := session.DB("ballots").C("polls")
-	p := NewPath(r.URL.Path)
-	if !p.HasID() {
-		respondErr(w, r, http.StatusMethodNotAllowed, "Cannot delete all polls.")
-		return
-	}
-	if err := c.RemoveId(bson.ObjectIdHex(p.ID)); err != nil {
-		respondErr(w, r, http.StatusInternalServerError, "failed to delete poll", err)
-		return
-	}
-	respond(w, r, http.StatusOK, nil) // ok
+// PollStorage represents an abstraction to a storage
+// backend for Polls
+type PollStorage interface {
+	Get(p *Path) ([]*Poll, error)
+	Create(p *Poll) (string, error)
+	Delete(p *Path) error
 }
